refactor(validator): extract CLI logic from main.rs
